backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
jinja2==3.1.2
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI/ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.15.0
torch==2.1.0
transformers==4.35.0

# Market Data and Analysis
ccxt==4.1.75
finta==1.3  # Technical analysis indicators (alternative to ta-lib)
yfinance==0.2.28

# HTTP and API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Caching and Performance
redis==5.0.1
celery==5.3.4
apscheduler==3.10.4

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
psutil==5.9.8

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Environment and Configuration
python-dotenv==1.0.0
